Silently ignore saved searches current user doesn't have rights to view
This could be a useful feature that for a dashboard used by multiple
users, some searches in the dashboard are only shown to ones with proper
rights. In that case, showing error message could confuse people.

Previously the main rights check is done implicitly via:

    ref( $SearchArg = $search->Content ) eq 'HASH'

Here we switch to CurrentUserHasRight('display') to make the logic a bit
more obvious.