Switch key/value pair if value is custom field value and key is not for Pg 9
For criteria that custom field value as the value, Pg 9 tries to cast
*all* ObjectCustomFieldValues content to the type of key instead of just
the ones bound to the specified custom field, e.g.

        LastUpdated > CF.{Beta Date}

If there is an ObjectCustomFieldValue row of which content doesn't look
like a datetime(e.g. a URL text for custom field "URL"), the cast will
fail.

This commit gets around this issue by switching key/value pair to
compare using text intead.